Facebook (#69)
* init Facebook bot

* init Facebook handle files

* extract handleMessage

* extract verifyRequestSignature

* refactor api

* rename GraphAPI

* add buttons

* fix error message

* extract formatButtons

* add send urban-img to facebook

* add logs

* format

* add quick_replies

* add audio video file

* add phoneNumber to button

* return button if type is specified

* init personaId

* change request to requestPromise

* move all file to ts and add facebook options

* rename env variables

* export types

* remove as

* update peerDependencies

* remove constructor body

* fix review

* remove logs

+++ packages/urban-bot/src/ManagerBot/ManagerBot.ts
@@ -116,10 +116,22 @@ export class ManagerBot<Bot extends UrbanBot = any> {
     }
 
     updateMessage(message: UrbanExistingMessage<BotMetaByBot<Bot>['MessageMeta']>) {
+        if (this.bot.updateMessage === undefined) {
+            throw new Error(
+                `'${this.bot.type}' doesn't support updating message. Provide isNewMessageEveryRender prop to Root component`,
+            );
+        }
+
         return this.bot.updateMessage(message);
     }
 
     deleteMessage(message: UrbanExistingMessage<BotMetaByBot<Bot>['MessageMeta']>) {
+        if (this.bot.deleteMessage === undefined) {
+            throw new Error(
+                `'${this.bot.type}' doesn't support deleting message. Provide isNewMessageEveryRender prop to Root component`,
+            );
+        }
+
         return this.bot.deleteMessage(message);
     }
 }
